update Admin Dash poard  Role  and update .gitignore

+++ MabeetApi/MabeetApi/Controllers/Admin/AdminAccommodationsController.cs
@@ -1,6 +1,8 @@
 ﻿using MabeetApi.DTOs.Admin.Accommodations;
 using MabeetApi.Services.Admin.Accommodations;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace MabeetApi.Controllers.Admin
 {
@@ -16,17 +18,27 @@ namespace MabeetApi.Controllers.Admin
         }
 
         // GET: api/admin/accommodations
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            var role = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (role != "Admin")
+                return Forbid();
             var accommodations = await _adminAccommodationService.GetAllAsync();
             return Ok(accommodations);
         }
 
         // GET: api/admin/accommodations/{id}
+        [Authorize]
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+            var role = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (role != "Admin")
+                return Forbid();
             var accommodation = await _adminAccommodationService.GetByIdAsync(id);
             if (accommodation == null)
                 return NotFound();
@@ -35,9 +47,14 @@ namespace MabeetApi.Controllers.Admin
         }
 
         // PUT: api/admin/accommodations/{id}/status
+        [Authorize]
         [HttpPut("{id}/status")]
         public async Task<IActionResult> UpdateStatus(int id, [FromBody] UpdateAccommodationStatusDto dto)
         {
+            var role = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (role != "Admin")
+                return Forbid();
             var result = await _adminAccommodationService.UpdateStatusAsync(id, dto);
             if (!result)
                 return NotFound();
@@ -46,9 +63,14 @@ namespace MabeetApi.Controllers.Admin
         }
 
         // DELETE: api/admin/accommodations/{id}
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var role = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (role != "Admin")
+                return Forbid();
             var result = await _adminAccommodationService.DeleteAsync(id);
             if (!result)
                 return NotFound();
